+++ Sources/Errors.swift
@@ -1,14 +1,19 @@
 public class TemplateDoesNotExist: Error, CustomStringConvertible {
   let templateNames: [String]
-  let loader: Loader
+  let loader: Loader?
 
-  public init(templateNames: [String], loader: Loader) {
+  public init(templateNames: [String], loader: Loader? = nil) {
     self.templateNames = templateNames
     self.loader = loader
   }
 
   public var description: String {
     let templates = templateNames.joined(separator: ", ")
-    return "Template named `\(templates)` does not exist in loader \(loader)"
+
+    if let loader = loader {
+      return "Template named `\(templates)` does not exist in loader \(loader)"
+    }
+
+    return "Template named `\(templates)` does not exist. No loaders found"
   }
 }