+++ Sources/Include.swift
@@ -19,16 +19,15 @@ class IncludeNode : NodeType {
   }
 
   func render(_ context: Context) throws -> String {
-    guard let loader = context["loader"] as? Loader else {
-      throw TemplateSyntaxError("Template loader not in context")
-    }
-
     guard let templateName = try self.templateName.resolve(context) as? String else {
       throw TemplateSyntaxError("'\(self.templateName)' could not be resolved as a string")
     }
 
-    let template = try loader.loadTemplate(name: templateName)
-    return try template.render(context)
+    let template = try context.environment.loadTemplate(name: templateName)
+
+    return try context.push {
+      return try template.render(context)
+    }
   }
 }