+++ Tests/StencilTests/TemplateSpec.swift
@@ -5,16 +5,14 @@ import Stencil
 func testTemplate() {
   describe("Template") {
     $0.it("can render a template from a string") {
-      let context = Context(dictionary: [ "name": "Kyle" ])
       let template = Template(templateString: "Hello World")
-      let result = try template.render(context)
+      let result = try template.render([ "name": "Kyle" ])
       try expect(result) == "Hello World"
     }
 
     $0.it("can render a template from a string literal") {
-        let context = Context(dictionary: [ "name": "Kyle" ])
         let template: Template = "Hello World"
-        let result = try template.render(context)
+        let result = try template.render([ "name": "Kyle" ])
         try expect(result) == "Hello World"
     }
   }